Enhance settings interface; add GUI sliders for adjustable parameters and refactor settings data types to float

+++ core/settings.hpp
@@ -4,15 +4,15 @@
 struct Settings {
     float atomicFriction = 0.975f;
     float forceStrength = 0.2f;
-    float forceMinDist = 15;
-    float forceMaxDist = 50;
-    float bondDist = 22;
+    float forceMinDist = 15.0f;
+    float forceMaxDist = 50.0f;
+    float bondDist = 22.0f;
     float decayChance = 0.04f;
-    int decayProtonThreshold = 84;
-    int fusionProtonMax = 70;
+    float decayProtonThreshold = 84.0f;
+    float fusionProtonMax = 70.0f;
     float temp = 1.0f;
     float electronOrbitSpeed = 0.015f;
-    int fontSize = 8;
+    float fontSize = 8.0f;
 };
 
 #endif
